+++ crates/rust-client/src/mock.rs
@@ -40,7 +40,7 @@ use crate::{
             block_header::BlockHeader as NodeBlockHeader,
             note::NoteSyncRecord,
             requests::SyncStateRequest,
-            responses::{NullifierUpdate, SyncStateResponse},
+            responses::{NullifierUpdate, SyncNoteResponse, SyncStateResponse},
         },
         AccountDetails, NodeRpcClient, NodeRpcClientEndpoint, NoteDetails, NoteInclusionDetails,
         RpcError, StateSyncInfo,
@@ -86,18 +86,28 @@ pub struct MockRpcApi {
     pub notes: BTreeMap<NoteId, InputNote>,
     pub mmr: Mmr,
     pub blocks: Vec<BlockHeader>,
+    pub sync_note_request: SyncNoteResponse,
 }
 
 impl Default for MockRpcApi {
     fn default() -> Self {
         let (genesis_block, state_sync_requests, notes, mmr, blocks) =
             generate_state_sync_mock_requests();
+
+        let sync_note_request = SyncNoteResponse {
+            chain_tip: 10,
+            notes: vec![],
+            block_header: Some(BlockHeader::mock(1, None, None, &[]).into()),
+            mmr_path: Some(Default::default()),
+        };
+
         Self {
             state_sync_requests,
             genesis_block,
             notes,
             mmr,
             blocks,
+            sync_note_request,
         }
     }
 }
@@ -109,6 +119,16 @@ impl MockRpcApi {
 }
 
 impl NodeRpcClient for MockRpcApi {
+    async fn sync_notes(
+        &mut self,
+        _block_num: u32,
+        _note_tags: &[NoteTag],
+    ) -> Result<crate::rpc::NoteSyncInfo, RpcError> {
+        let response = &self.sync_note_request;
+        let response = Response::new(response.clone());
+        response.into_inner().try_into()
+    }
+
     /// Executes the specified sync state request and returns the response.
     async fn sync_state(
         &mut self,