+++ crates/rust-client/src/rpc/tonic_client/mod.rs
@@ -7,9 +7,9 @@ use std::time::Duration;
 use generated::{
     requests::{
         GetAccountDetailsRequest, GetBlockHeaderByNumberRequest, GetNotesByIdRequest,
-        SubmitProvenTransactionRequest, SyncStateRequest,
+        SubmitProvenTransactionRequest, SyncNoteRequest, SyncStateRequest,
     },
-    responses::SyncStateResponse,
+    responses::{SyncNoteResponse, SyncStateResponse},
     rpc::api_client::ApiClient,
 };
 use miden_objects::{
@@ -26,7 +26,8 @@ use tracing::info;
 
 use super::{
     AccountDetails, AccountUpdateSummary, CommittedNote, NodeRpcClient, NodeRpcClientEndpoint,
-    NoteDetails, NoteInclusionDetails, NullifierUpdate, StateSyncInfo, TransactionUpdate,
+    NoteDetails, NoteInclusionDetails, NoteSyncInfo, NullifierUpdate, StateSyncInfo,
+    TransactionUpdate,
 };
 use crate::{config::RpcConfig, rpc::RpcError};
 
@@ -270,6 +271,74 @@ impl NodeRpcClient for TonicRpcClient {
         } else {
             Ok(AccountDetails::OffChain(account_id, update_summary))
         }
+    }
+
+    async fn sync_notes(
+        &mut self,
+        block_num: u32,
+        note_tags: &[NoteTag],
+    ) -> Result<super::NoteSyncInfo, RpcError> {
+        let note_tags = note_tags.iter().map(|&note_tag| note_tag.into()).collect();
+
+        let request = SyncNoteRequest { block_num, note_tags };
+
+        let rpc_api = self.rpc_api().await?;
+
+        let response = rpc_api.sync_notes(request).await.map_err(|err| {
+            RpcError::RequestError(NodeRpcClientEndpoint::SyncNotes.to_string(), err.to_string())
+        })?;
+
+        response.into_inner().try_into()
+    }
+}
+
+// NOTE SYNC INFO CONVERSION
+// ================================================================================================
+
+impl TryFrom<SyncNoteResponse> for NoteSyncInfo {
+    type Error = RpcError;
+
+    fn try_from(value: SyncNoteResponse) -> Result<Self, Self::Error> {
+        let chain_tip = value.chain_tip;
+
+        // Validate and convert block header
+        let block_header = value
+            .block_header
+            .ok_or(RpcError::ExpectedFieldMissing("BlockHeader".into()))?
+            .try_into()?;
+
+        let mmr_path = value
+            .mmr_path
+            .ok_or(RpcError::ExpectedFieldMissing("MmrPath".into()))?
+            .try_into()?;
+
+        // Validate and convert account note inclusions into an (AccountId, Digest) tuple
+        let mut notes = vec![];
+        for note in value.notes {
+            let note_id: Digest = note
+                .note_id
+                .ok_or(RpcError::ExpectedFieldMissing("Notes.Id".into()))?
+                .try_into()?;
+
+            let note_id: NoteId = note_id.into();
+
+            let merkle_path = note
+                .merkle_path
+                .ok_or(RpcError::ExpectedFieldMissing("Notes.MerklePath".into()))?
+                .try_into()?;
+
+            let metadata = note
+                .metadata
+                .ok_or(RpcError::ExpectedFieldMissing("Metadata".into()))?
+                .try_into()?;
+
+            let committed_note =
+                CommittedNote::new(note_id, note.note_index, merkle_path, metadata);
+
+            notes.push(committed_note);
+        }
+
+        Ok(NoteSyncInfo { chain_tip, block_header, mmr_path, notes })
     }
 }