+++ tests/integration/swap_transactions_tests.rs
@@ -360,8 +360,8 @@ async fn test_swap_offchain() {
     client2
         .import_note(NoteFile::NoteDetails {
             details: exported_note.into(),
+            after_block_num: client1.get_sync_height().unwrap(),
             tag: Some(tag),
-            after_block_num: 0,
         })
         .await
         .unwrap();